Super-commit: Documenation, bugfixes, sign-up, avatars...

+++ api.php
@@ -1,10 +1,12 @@
 <?php
 
-// Commentia API
-// This should be the ONLY entry point directly from a website.
-// Anything after the $_SESSION['member_is_logged_in'] check can only be executed ince authenticated.
-// Author: Alexander Gilburg
-// Last updated: 15th of July 2016
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+// Commentia API                                                                                      //
+// This should be the ONLY entry point directly from a website.                                       //
+// Anything after the $_SESSION['member_is_logged_in'] check can only be executed ince authenticated. //
+// Author: Alexander Gilburg                                                                          //
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 
 session_start();
 
@@ -31,16 +33,12 @@ use Commentia\Roles\Roles;
 
 $roles = new Roles();
 
-if (!isset($_SESSION['member_is_logged_in'])) {
-  $_SESSION['member_is_logged_in'] = false;
-}
-
 if (isset($_POST['action'])) {
     $action = $_POST['action'];
 
     if (($action === 'loginMember')
-  && isset($_POST['username'])
-  && isset($_POST['password'])) {
+    && isset($_POST['username'])
+    && isset($_POST['password'])) {
         $username = $_POST['username'];
         $password = $_POST['password'];
         $commentia->loginMember($username, $password);
@@ -48,6 +46,15 @@ if (isset($_POST['action'])) {
 
     if (($action === 'logoutMember')) {
         $commentia->logoutMember();
+    }
+
+    if ($action === 'signUpMember') {
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $retyped_password = $_POST['retyped_password'];
+        $email = $_POST['email'];
+        $avatar_file = $_POST['avatar_file'];
+        $commentia->signUpMember($username, $password, $retyped_password, $email, $avatar_file);
     }
 }
 
@@ -57,7 +64,8 @@ if (isset($_SESSION['pageid'])
     $action = $_GET['action'];
 
     if ($action === 'display') {
-        echo $commentia->displayComments();
+        $is_ajax_request = true;
+        echo $commentia->displayComments($is_ajax_request);
     }
 
     if ($action === 'getPhrase'
@@ -83,8 +91,8 @@ if (isset($_SESSION['pageid'])
     $action = $_POST['action'];
 
     if ((($action === 'reply')
-  || ($action === 'postNewComment'))
-  && isset($_POST['content'])) {
+    || ($action === 'postNewComment'))
+    && isset($_POST['content'])) {
         $content = $_POST['content'];
         $reply_path = $_POST['reply_path'];
 
@@ -92,10 +100,10 @@ if (isset($_SESSION['pageid'])
     }
 
     if (($roles->memberHasUsername($commentia->getCommentData($_POST['ucid'], $_POST['reply_path'], 'creator_username'))
-  || $roles->memberIsAdmin())
-  && ($action === 'edit')
-  && isset($_POST['content'])
-  && isset($_POST['ucid'])) {
+    || $roles->memberIsAdmin())
+    && ($action === 'edit')
+    && isset($_POST['content'])
+    && isset($_POST['ucid'])) {
         $content = $_POST['content'];
         $ucid = $_POST['ucid'];
         $reply_path = $_POST['reply_path'];
@@ -103,9 +111,9 @@ if (isset($_SESSION['pageid'])
     }
 
     if (($roles->memberHasUsername($commentia->getCommentData($_POST['ucid'], $_POST['reply_path'], 'creator_username'))
-  || $roles->memberIsAdmin())
-  && ($action === 'delete')
-  && isset($_POST['ucid'])) {
+    || $roles->memberIsAdmin())
+    && ($action === 'delete')
+    && isset($_POST['ucid'])) {
         $ucid = $_POST['ucid'];
         $reply_path = $_POST['reply_path'];
         $commentia->deleteComment($ucid, $reply_path);
